⬆️♻️chore(all-the-app): Updating the packages and refactoring
I updated all the packages of the frontend version and changed a bit of code to make it work again.
I mainly changed:
	- React router DOM components
	- MaterialUI
	- Modification du fichier routes.js pour adapter aux nouvelles balises (navigate, Routes, Route, etc)
MaterialUI is now changed to MUI and uses "emotion"

I still have some css to change and some warning to resolve but the app is working more or less correctly.